requirements.txt
google-generativeai
streamlit
pandas
openpyxl
requests
numpy
scikit-learn
